+++ mne/io/edf/edf.py
@@ -18,7 +18,7 @@ import re
 import numpy as np
 
 from ...utils import verbose, logger, warn
-from ..utils import _blk_read_lims
+from ..utils import _blk_read_lims, _deprecate_montage
 from ..base import BaseRaw
 from ..meas_info import _empty_info, _unique_channel_names, DATE_NONE
 from ..constants import FIFF
@@ -83,10 +83,7 @@ class RawEDF(BaseRaw):
     ----------
     input_fname : str
         Path to the EDF, EDF+ or BDF file.
-    montage : str | None | instance of Montage
-        Path or instance of montage containing electrode positions. If None,
-        sensor locations are (0,0,0). See the documentation of
-        :func:`mne.channels.read_montage` for more information.
+    %(montage_deprecated)s
     eog : list or tuple
         Names of channels or list of indices that should be designated EOG
         channels. Values should correspond to the electrodes in the file.
@@ -185,9 +182,8 @@ class RawEDF(BaseRaw):
 
         self.set_annotations(Annotations(onset=onset, duration=duration,
                                          description=desc, orig_time=None))
-        if montage is not None:
-            # XXX: set_montage(montage=None) will change all locations to NaN
-            self.set_montage(montage)
+
+        _deprecate_montage(self, "read_raw_edf", montage)
 
     @verbose
     def _read_segment_file(self, data, idx, fi, start, stop, cals, mult):
@@ -211,10 +207,7 @@ class RawGDF(BaseRaw):
     ----------
     input_fname : str
         Path to the GDF file.
-    montage : str | None | instance of Montage
-        Path or instance of montage containing electrode positions. If None,
-        sensor locations are (0,0,0). See the documentation of
-        :func:`mne.channels.read_montage` for more information.
+    %(montage_deprecated)s
     eog : list or tuple
         Names of channels or list of indices that should be designated EOG
         channels. Values should correspond to the electrodes in the file.
@@ -269,8 +262,8 @@ class RawGDF(BaseRaw):
 
         self.set_annotations(Annotations(onset=onset, duration=duration,
                                          description=desc, orig_time=None))
-        if montage is not None:
-            self.set_montage(montage)
+
+        _deprecate_montage(self, "read_raw_gdf", montage)
 
     @verbose
     def _read_segment_file(self, data, idx, fi, start, stop, cals, mult):
@@ -1153,7 +1146,7 @@ def _find_tal_idx(ch_names):
 
 
 @fill_doc
-def read_raw_edf(input_fname, montage=None, eog=None, misc=None,
+def read_raw_edf(input_fname, montage='deprecated', eog=None, misc=None,
                  stim_channel='auto', exclude=(), preload=False, verbose=None):
     """Reader function for EDF or EDF+ files.
 
@@ -1161,10 +1154,7 @@ def read_raw_edf(input_fname, montage=None, eog=None, misc=None,
     ----------
     input_fname : str
         Path to the EDF or EDF+ file.
-    montage : str | None | instance of Montage
-        Path or instance of montage containing electrode positions. If None,
-        sensor locations are (0,0,0). See the documentation of
-        :func:`mne.channels.read_montage` for more information.
+    %(montage_deprecated)s
     eog : list or tuple
         Names of channels or list of indices that should be designated EOG
         channels. Values should correspond to the electrodes in the file.
@@ -1234,7 +1224,7 @@ def read_raw_edf(input_fname, montage=None, eog=None, misc=None,
 
 
 @fill_doc
-def read_raw_bdf(input_fname, montage=None, eog=None, misc=None,
+def read_raw_bdf(input_fname, montage='deprecated', eog=None, misc=None,
                  stim_channel='auto', exclude=(), preload=False, verbose=None):
     """Reader function for BDF files.
 
@@ -1242,10 +1232,7 @@ def read_raw_bdf(input_fname, montage=None, eog=None, misc=None,
     ----------
     input_fname : str
         Path to the BDF file.
-    montage : str | None | instance of Montage
-        Path or instance of montage containing electrode positions. If None,
-        sensor locations are (0,0,0). See the documentation of
-        :func:`mne.channels.read_montage` for more information.
+    %(montage_deprecated)s
     eog : list or tuple
         Names of channels or list of indices that should be designated EOG
         channels. Values should correspond to the electrodes in the file.
@@ -1325,7 +1312,7 @@ def read_raw_bdf(input_fname, montage=None, eog=None, misc=None,
 
 
 @fill_doc
-def read_raw_gdf(input_fname, montage=None, eog=None, misc=None,
+def read_raw_gdf(input_fname, montage='deprecated', eog=None, misc=None,
                  stim_channel='auto', exclude=(), preload=False, verbose=None):
     """Reader function for GDF files.