+++ mne/io/egi/egi.py
@@ -11,7 +11,7 @@ import numpy as np
 from .egimff import _read_raw_egi_mff
 from .events import _combine_triggers
 from ..base import BaseRaw
-from ..utils import _read_segments_file, _create_chs
+from ..utils import _read_segments_file, _create_chs, _deprecate_montage
 from ..meas_info import _empty_info
 from ..constants import FIFF
 from ...utils import verbose, logger, warn
@@ -88,7 +88,7 @@ def _read_events(fid, info):
 
 
 @verbose
-def read_raw_egi(input_fname, montage=None, eog=None, misc=None,
+def read_raw_egi(input_fname, montage='deprecated', eog=None, misc=None,
                  include=None, exclude=None, preload=False,
                  channel_naming='E%d', verbose=None):
     """Read EGI simple binary as raw object.
@@ -98,10 +98,7 @@ def read_raw_egi(input_fname, montage=None, eog=None, misc=None,
     input_fname : str
         Path to the raw file. Files with an extension .mff are automatically
         considered to be EGI's native MFF format files.
-    montage : str | None | instance of Montage
-        Path or instance of montage containing electrode positions.
-        If None, sensor locations are (0,0,0). See the documentation of
-        :func:`mne.channels.read_montage` for more information.
+    %(montage_deprecated)s
     eog : list or tuple
         Names of channels or list of indices that should be designated
         EOG channels. Default is None.
@@ -163,7 +160,7 @@ class RawEGI(BaseRaw):
     """Raw object from EGI simple binary file."""
 
     @verbose
-    def __init__(self, input_fname, montage=None, eog=None, misc=None,
+    def __init__(self, input_fname, montage='deprecated', eog=None, misc=None,
                  include=None, exclude=None, preload=False,
                  channel_naming='E%d', verbose=None):  # noqa: D102
         if eog is None:
@@ -261,8 +258,7 @@ class RawEGI(BaseRaw):
             filenames=[input_fname], last_samps=[egi_info['n_samples'] - 1],
             raw_extras=[egi_info], verbose=verbose)
 
-        if montage is not None:
-            self.set_montage(montage)
+        _deprecate_montage(self, "read_raw_egi", montage)
 
     def _read_segment_file(self, data, idx, fi, start, stop, cals, mult):
         """Read a segment of data from a file."""